refactor(types): вынести коллекционные свойства TypeRegistry в отдельный индекс
Пять индексов коллекционных свойств — типы элементов по умолчанию, признаки
Для Каждого и индексатора вместе с их текстовыми описаниями — и работающие
только с ними операции переехали в CollectionTraitsIndex. TypeRegistry
делегирует к нему; публичный контракт реестра не изменился.

Повод — java:S104: файл перешагнул порог в 750 строк кода (751), причём
перешагнул именно правкой этой ветки. После выноса — 703.

Заодно unregisterUserType чистит все пять индексов вместо двух: раньше
supportsIndexAccess и описания остались бы висеть. Практической разницы нет —
USER-типам их никто не выставляет, — но remove() обещает полную очистку.
