+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/registry/TypeRegistry.java
@@ -184,22 +184,11 @@ public class TypeRegistry {
   private final Map<FileType, Map<TypeRef, List<Supplier<List<SignatureDescriptor>>>>> constructorSources = perFileType();
 
   /**
-   * Тип ↔ типы элементов «по умолчанию» для коллекции. Заполняется из
-   * {@link TypePackProvider.TypeDecl#defaultElementTypes()} при регистрации.
-   * Источник истины — bsl-context ({@code ContextCollection.collectionElementTypes()})
-   * или builtin JSON. Используется инференсером для прокидывания element-типа
-   * на TypeSet, чтобы {@code Для Каждого X Из Коллекция} давал X нужного типа
-   * без аннотаций пользователя.
+   * Коллекционные свойства типов: типы элементов, признаки {@code Для Каждого}
+   * и индексатора с описаниями. Вынесены в отдельный индекс — реестр только
+   * делегирует к нему и канонизирует типы элементов через {@link #resolve(String)}.
    */
-  private final Map<TypeRef, List<TypeRef>> defaultElementTypes = new ConcurrentHashMap<>();
-  /** Тип ↔ {@code supportsForEach} в разрезе языка ({@code true} — обход {@code Для Каждого} разрешён). */
-  private final Map<FileType, Map<TypeRef, Boolean>> supportsForEach = perFileType();
-  /** Тип ↔ {@code supportsIndexAccess} в разрезе языка ({@code true} — индексатор {@code [...]} разрешён). */
-  private final Map<FileType, Map<TypeRef, Boolean>> supportsIndexAccess = perFileType();
-  /** Тип ↔ текстовое описание обхода {@code Для Каждого} из синтакс-помощника, в разрезе языка. */
-  private final Map<FileType, Map<TypeRef, BilingualString>> forEachDescriptions = perFileType();
-  /** Тип ↔ текстовое описание индексатора {@code [...]} из синтакс-помощника, в разрезе языка. */
-  private final Map<FileType, Map<TypeRef, BilingualString>> indexAccessDescriptions = perFileType();
+  private final CollectionTraitsIndex collectionTraits = new CollectionTraitsIndex();
   /**
    * Тип ↔ имена generic-плейсхолдеров (без угловых скобок). Заполняется
    * платформенным провайдером из {@link TypePackProvider.TypeDecl#typeParameters()}.
@@ -1222,8 +1211,7 @@ public class TypeRegistry {
     membersEpoch.incrementAndGet();
     visibleTypes.values().forEach(typed -> typed.remove(ref));
     aliasIndex.remove(qualifiedName.toLowerCase(Locale.ROOT));
-    supportsForEach.values().forEach(byRef -> byRef.remove(ref));
-    defaultElementTypes.remove(ref);
+    collectionTraits.remove(ref);
   }
 
   private void registerPack(TypePackProvider.TypeDecl decl, FileType fileType) {
@@ -1236,7 +1224,8 @@ public class TypeRegistry {
     registerPackDescriptions(decl, ref, fileType);
     registerPackMetadata(decl, ref, fileType);
     registerPackCallables(decl, ref, fileType);
-    registerPackCollectionTraits(decl, ref, fileType);
+    collectionTraits.registerPack(decl, ref, fileType);
+    registerPackTypeParameters(decl, ref);
     if (!decl.name().isEmpty()) {
       displayNames.putIfAbsent(ref, decl.name());
     }
@@ -1293,23 +1282,8 @@ public class TypeRegistry {
     }
   }
 
-  /** Коллекционные свойства пака: элементы по умолчанию, Для Каждого, индексатор, generic-параметры. */
-  private void registerPackCollectionTraits(TypePackProvider.TypeDecl decl, TypeRef ref, FileType fileType) {
-    if (decl.defaultElementTypes() != null && !decl.defaultElementTypes().isEmpty()) {
-      defaultElementTypes.put(ref, List.copyOf(decl.defaultElementTypes()));
-    }
-    if (decl.supportsForEach()) {
-      supportsForEach.get(fileType).put(ref, Boolean.TRUE);
-    }
-    if (decl.supportsIndexAccess()) {
-      supportsIndexAccess.get(fileType).put(ref, Boolean.TRUE);
-    }
-    if (!decl.forEachDescription().isEmpty()) {
-      forEachDescriptions.get(fileType).put(ref, decl.forEachDescription());
-    }
-    if (!decl.indexAccessDescription().isEmpty()) {
-      indexAccessDescriptions.get(fileType).put(ref, decl.indexAccessDescription());
-    }
+  /** Generic-параметры пака: имена плейсхолдеров в qualifiedName типа. */
+  private void registerPackTypeParameters(TypePackProvider.TypeDecl decl, TypeRef ref) {
     if (!decl.typeParameters().isEmpty()) {
       typeParameters.put(ref, List.copyOf(decl.typeParameters()));
     }
@@ -1344,22 +1318,13 @@ public class TypeRegistry {
    * используются как ключи в индексах членов).
    */
   public TypeSet getDefaultElementTypes(TypeRef ref) {
-    var raw = defaultElementTypes.get(ref);
-    if (raw == null || raw.isEmpty()) {
-      return TypeSet.EMPTY;
-    }
-    var canonical = new ArrayList<TypeRef>(raw.size());
-    for (var element : raw) {
-      var resolved = resolve(element.qualifiedName()).orElse(element);
-      canonical.add(resolved);
-    }
-    return TypeSet.of(canonical);
+    return collectionTraits.defaultElementTypes(ref, element -> resolve(element.qualifiedName()).orElse(element));
   }
 
   /**
    * Пометить пользовательский тип (OneScript-класс) как коллекцию, обходимую
    * через {@code Для Каждого} — или снять признак. Это императивный аналог
-   * {@link #registerPackCollectionTraits} для USER-типов, у которых нет
+   * регистрации пака для USER-типов, у которых нет
    * {@link TypePackProvider.TypeDecl}: источник истины — аннотация
    * {@code &Обходимое} на {@code ПриСозданииОбъекта} (см.
    * {@code OScriptIterable#isIterable}).
@@ -1378,11 +1343,7 @@ public class TypeRegistry {
    * @param fileType языковой скоуп, в котором действует признак.
    */
   public void setUserTypeIterable(TypeRef ref, boolean iterable, FileType fileType) {
-    if (iterable) {
-      supportsForEach.get(fileType).put(ref, Boolean.TRUE);
-    } else {
-      supportsForEach.get(fileType).remove(ref);
-    }
+    collectionTraits.setIterable(ref, iterable, fileType);
   }
 
   /**
@@ -1404,34 +1365,17 @@ public class TypeRegistry {
    * @param fileType языковой скоуп.
    */
   public void inheritCollectionTraits(TypeRef target, TypeRef source, FileType fileType) {
-    if (target.equals(source)) {
-      return;
-    }
-    var elements = defaultElementTypes.get(source);
-    if (elements != null && !elements.isEmpty()) {
-      defaultElementTypes.putIfAbsent(target, elements);
-    }
-    copyTrait(supportsForEach.get(fileType), target, source);
-    copyTrait(supportsIndexAccess.get(fileType), target, source);
-    copyTrait(forEachDescriptions.get(fileType), target, source);
-    copyTrait(indexAccessDescriptions.get(fileType), target, source);
-  }
-
-  private static <V> void copyTrait(Map<TypeRef, V> trait, TypeRef target, TypeRef source) {
-    var value = trait.get(source);
-    if (value != null) {
-      trait.putIfAbsent(target, value);
-    }
+    collectionTraits.inherit(target, source, fileType);
   }
 
   /** {@code true}, если у типа разрешён обход {@code Для Каждого} в данном языке файла. */
   public boolean supportsForEach(TypeRef ref, FileType fileType) {
-    return Boolean.TRUE.equals(supportsForEach.get(fileType).get(ref));
+    return collectionTraits.supportsForEach(ref, fileType);
   }
 
   /** {@code true}, если у типа разрешён индексатор {@code [...]} в данном языке файла. */
   public boolean supportsIndexAccess(TypeRef ref, FileType fileType) {
-    return Boolean.TRUE.equals(supportsIndexAccess.get(fileType).get(ref));
+    return collectionTraits.supportsIndexAccess(ref, fileType);
   }
 
   /**
@@ -1445,7 +1389,7 @@ public class TypeRegistry {
 
   /** Описание обхода в указанной локали (с fallback на другую) в данном языке файла. */
   public String getForEachDescription(TypeRef ref, FileType fileType, Language language) {
-    return forEachDescriptions.get(fileType).getOrDefault(ref, BilingualString.EMPTY).forLanguage(language);
+    return collectionTraits.forEachDescription(ref, fileType, language);
   }
 
   /**
@@ -1459,7 +1403,7 @@ public class TypeRegistry {
 
   /** Описание индексатора в указанной локали в данном языке файла. */
   public String getIndexAccessDescription(TypeRef ref, FileType fileType, Language language) {
-    return indexAccessDescriptions.get(fileType).getOrDefault(ref, BilingualString.EMPTY).forLanguage(language);
+    return collectionTraits.indexAccessDescription(ref, fileType, language);
   }
 
   /**
